refactor: 将初始化项目从设置页移到独立页面
- 新增 ProjectInit 页面，专门用于创建新 Hugo 站点
- 在侧边栏添加'初始化项目'入口
- 从 Settings 页面移除初始化项目表单，减少误触风险
